Update auto SPI for timestamp changes (#1457)
The 2019 FPGA image switched the output of auto SPI from plain bytes to a
sequence of 32-bit words (timestamp, then words with the byte values in the
least significant byte of each word).

In addition to changing the HAL and simulators to reflect this, add piecewise
integration support to wpilibc/wpilibj SPI to take advantage of the timestamps
and use it in the ADXRS450 gyro.

+++ hal/src/main/native/sim/mockdata/SPIDataInternal.h
@@ -24,7 +24,7 @@ class SPIData {
   int32_t Write(const uint8_t* dataToSend, int32_t sendSize);
   int32_t Transaction(const uint8_t* dataToSend, uint8_t* dataReceived,
                       int32_t size);
-  int32_t ReadAutoReceivedData(uint8_t* buffer, int32_t numToRead,
+  int32_t ReadAutoReceivedData(uint32_t* buffer, int32_t numToRead,
                                double timeout, int32_t* status);
 
   SimDataValue<HAL_Bool, MakeBoolean, GetInitializedName> initialized{false};
